requirements.txt
certifi==2024.2.2
chardet==5.2.0
charset-normalizer==3.3.2
decorator==4.4.2
idna==3.7
imageio==2.34.1
imageio-ffmpeg==0.5.0
moviepy==1.0.3
numpy==1.26.4
pillow==10.3.0
proglog==0.1.10
reportlab==4.2.0
requests==2.32.3
tqdm==4.66.4
urllib3==2.2.1
